Renamed multivariate_meanvar_score to mean_cov_score
