+++ skchange/change_detectors/moscore.py
@@ -80,37 +80,11 @@ def moscore_transform(
     score_f: Callable,
     score_init_f: Callable,
     bandwidth: int,
-) -> Tuple[list, np.ndarray]:
+) -> tuple[list, np.ndarray]:
     n = len(X)
-    # splits = np.arange(bandwidth - 1, n - bandwidth)
-    # starts = splits - bandwidth + 1
-    # ends = splits + bandwidth
-    # When accessing data, access as:
-    # X[starts[i]:(splits[i] + 1)]
-    # Get 'bandwidth' elements before the split.
-    # X[splits[i]:(ends[i] + 1)]
-    # Get 'bandwidth + 1'elements after the split:
-    # Or: X[splits[i]:ends[i]], but this excludes last data point. 
-
-    # Split up into equal sized windows:
-    # Each window: [start[i], split[i] - 1] (inclusive)
-    #              [split[i], end[i]] (inclusive)
-    # Total of 'bandwidth' elements in each window.
-
-    # Shifted 'splits' up by one:
-    splits = np.arange(bandwidth, n - bandwidth + 1)
-    # Exactly the same 'starts' as above:
-    starts = splits - bandwidth
-    # Exactly the same 'ends' as above:
-    ends = splits + bandwidth - 1
-
-    # Access as:
-    # X[starts[i]:splits[i]]
-    # Get 'bandwidth' elements before the split.
-    # X[splits[i]:(ends[i] + 1)]
-    # Get 'bandwidth' elements after the split.
-    # Includes the last data point in the second window.
-
+    splits = np.arange(bandwidth - 1, n - bandwidth)
+    starts = splits - bandwidth + 1
+    ends = splits + bandwidth
     params = score_init_f(X)
     scores = np.zeros(n)
     scores[splits] = score_f(params, starts, ends, splits)
@@ -195,16 +169,14 @@ class Moscore(BaseSeriesAnnotator):
         min_detection_interval: int = 1,
         fmt: str = "sparse",
         labels: str = "int_label",
-        require_jit: bool = True,
     ):
         self.score = score
         self.bandwidth = bandwidth
         self.threshold_scale = threshold_scale  # Just holds the input value.
         self.level = level
         self.min_detection_interval = min_detection_interval
-        self.require_jit = require_jit
         super().__init__(fmt=fmt, labels=labels)
-        self.score_f, self.score_init_f = score_factory(self.score, self.require_jit)
+        self.score_f, self.score_init_f = score_factory(self.score)
 
         check_larger_than(1, self.bandwidth, "bandwidth")
         check_larger_than(0, threshold_scale, "threshold_scale", allow_none=True)
@@ -231,20 +203,12 @@ class Moscore(BaseSeriesAnnotator):
         X : pd.DataFrame
             Training data to tune the threshold on.
         """
-        if self.require_jit:
-            scores = moscore_transform(
-                X.values,
-                self.score_f,
-                self.score_init_f,
-                self.bandwidth,
-            )
-        else:
-            scores = moscore_transform_nojit(
-                X.values,
-                self.score_f,
-                self.score_init_f,
-                self.bandwidth,
-            )
+        scores = moscore_transform(
+            X.values,
+            self.score_f,
+            self.score_init_f,
+            self.bandwidth,
+        )
         tuned_threshold = np.quantile(scores, 1 - self.level)
         return tuned_threshold
 
@@ -349,20 +313,12 @@ class Moscore(BaseSeriesAnnotator):
             min_length=2 * self.bandwidth,
             min_length_name="2*bandwidth",
         )
-        if self.require_jit:
-            scores = moscore_transform(
-                X.values,
-                self.score_f,
-                self.score_init_f,
-                self.bandwidth,
-            )
-        else:
-            scores = moscore_transform_nojit(
-                X.values,
-                self.score_f,
-                self.score_init_f,
-                self.bandwidth,
-            )
+        scores = moscore_transform(
+            X.values,
+            self.score_f,
+            self.score_init_f,
+            self.bandwidth,
+        )
         self.changepoints = get_moscore_changepoints(
             scores, self.threshold_, self.min_detection_interval
         )
